+++ src/store/reducers/async/middleware.js
@@ -1,4 +1,18 @@
 import {batch} from "react-redux";
+import {load_all_data} from "../geoJsonReducer";
+
+export const fetchFeatures = () => {
+  return (dispatch) => {
+
+    fetch("data.json")
+      .then(response => response.json())
+      .then(data => {
+        dispatch(load_all_data(data.features))
+      }, (error) => {
+        console.error(error);
+      });
+  }
+}
 
 export const fetchAllIds = (tabId) => {
