+++ src/store/reducers/geoJsonReducer.js
@@ -2,7 +2,7 @@ const initState = {
   features: []
 }
 
-const dataReducer = (state = initState, action) => {
+const geoJsonReducer = (state = initState, action) => {
   switch (action.type) {
     case "LOAD_ALL_DATA":
       return { ...state, features: [...action.payload]}
@@ -13,4 +13,4 @@ const dataReducer = (state = initState, action) => {
 
 export const load_all_data = (payload) => ({type: "LOAD_ALL_DATA", payload})
 
-export default dataReducer
+export default geoJsonReducer
